requirements.txt
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
et-xmlfile==1.1.0
fastapi==0.86.0
h11==0.14.0
httptools==0.5.0
idna==3.4
numpy==1.23.4
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.1
pandas-stubs==1.5.1.221024
pydantic==1.10.2
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tqdm==4.64.1
twilio==7.15.2
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
